Move ratelimit to model dir
