+++ forecast_solar/models.py
@@ -5,6 +5,8 @@ from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from typing import Any
 
+from aiohttp import ClientResponse
+
 
 @dataclass
 class Estimate:
@@ -178,3 +180,31 @@ class Estimate:
             },
             api_timezone=data["message"]["info"]["timezone"],
         )
+
+
+@dataclass
+class Ratelimit:
+    """Information about the current rate limit."""
+
+    call_limit: int
+    remaining_calls: int
+    period: int
+    retry_at: datetime | None
+
+    @classmethod
+    def from_response(cls, response: ClientResponse) -> Ratelimit:
+        """Initialize rate limit object from response."""
+        # The documented headers do not match the returned headers
+        # https://doc.forecast.solar/doku.php?id=api#headers
+        limit = int(response.headers["X-Ratelimit-Limit"])
+        period = int(response.headers["X-Ratelimit-Period"])
+
+        # Remaining is not there if we exceeded limit
+        remaining = int(response.headers.get("X-Ratelimit-Remaining", 0))
+
+        if "X-Ratelimit-Retry-At" in response.headers:
+            retry_at = datetime.fromisoformat(response.headers["X-Ratelimit-Retry-At"])
+        else:
+            retry_at = None
+
+        return cls(limit, period, remaining, retry_at)
